Refactor ProcessWords_ShouldReturnWordsInLowerCase test, create folders Layouter and Spiral
