+++ TagsCloudContainer/WorkWithWords/WordGenerator.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Drawing;
+using TagsCloudContainer.Layouter;
 
 namespace TagsCloudContainer.WorkWithWords
 {
