+++ TagsCloudContainerTests/SpiralTest.cs
@@ -3,6 +3,7 @@ using System.Drawing;
 using FluentAssertions;
 using NUnit.Framework;
 using TagsCloudContainer;
+using TagsCloudContainer.Spiral;
 
 namespace TagsCloudContainerTests
 {
